✨ #267 - Adicionado botão na tela de Politica de privacidade
