+++ src/pages/Perfil/PoliticaDePrivacidade/styles.js
@@ -1,0 +1,27 @@
+import {StyleSheet} from 'react-native';
+
+const estilos = StyleSheet.create({
+  titulo: { fontSize: 23, marginTop: 30, textAlign: 'center' },
+  conteudoTexto: { marginHorizontal: 16, marginBottom: 50 },
+  texto: {
+    color: '#000000',
+    opacity: 0.6,
+    fontSize: 14,
+    marginBottom: 8,
+    marginTop: 8,
+    textAlign: 'justify'
+  },
+  button: {
+    alignItems: 'center',
+    backgroundColor: '#4CAF50',
+    padding: 10,
+    marginBottom: 28,
+    marginHorizontal: 16
+  },
+  texto_button: {
+    color: '#FFFFFF',
+    fontSize: 14
+  }
+});
+
+export default estilos;
